% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatter-AllGenerics.R,
%   R/splatter-SplatParams-methods.R, R/splatter-params-functions.R
\name{expandParams}
\alias{expandParams}
\alias{expandParams,SplatParams-method}
\alias{paramsExpander}
\title{Expand parameters}
\usage{
expandParams(object, ...)

\S4method{expandParams}{SplatParams}(object)

paramsExpander(object, vectors, n)
}
\arguments{
\item{object}{object to expand.}

\item{...}{additional arguments.}

\item{vectors}{names of vector parameters to expand}

\item{n}{number of times to repeat each parameter}
}
\value{
Expanded object.
}
\description{
Expand the parameters that can be vectors so that they are the same length as
the number of groups. Work is done by \code{\link{paramsExpander}} called
from each method. Expansions are stored using
\code{\link{setParamsUnchecked}}.
}
\keyword{internal}
