% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatter-params-functions.R
\name{getParams}
\alias{getParams}
\title{Get parameters}
\usage{
getParams(params, names)
}
\arguments{
\item{params}{Params object to get values from.}

\item{names}{vector of names of the parameters to get.}
}
\value{
List with the values of the selected parameters.
}
\description{
Get multiple parameter values from a Params object.
}
\examples{
params <- newSimpleParams()
getParams(params, c("nGenes", "nCells", "mean.rate"))
}
