% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSTMseq-package.R
\docType{_PACKAGE}
\name{scSTMseq-package}
\alias{scSTMseq-package}
\title{Single Cell Structural Topic Model Sequencing}
\description{
This package implements the Structural Topic Model, adapted to single cell RNAseq data,
a general approach to including cell-level metadata within mixed-membership topic
models. This package is built upon package stm, and borrowed functions from Splatter
}
\details{
Functions to simulate single cell data: \code{\link{SimPairedData}}

Functions to fit the model: \code{\link{scSTM}} \code{\link{selectModel}}
\code{\link{selectModel_parallel}}

Functions to summarize a model: \code{\link{labelTopics}}
\code{\link{summary.scSTM}}

Functions for Post-Estimation: \code{\link{estimateEffect}}
\code{\link{topicCorr}}

Plotting Functions: \code{\link{plot.scSTM}} \code{\link{structure_plot}}
\code{\link{plot.estimateEffect}}
}
\author{
Author: Euphy Wu, Didong Li, Naim Rashid
}
\keyword{_PACKAGE}
