% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSTMseq-simulate.R
\name{scSTMseqSimBatchEffects}
\alias{scSTMseqSimBatchEffects}
\title{Simulate gene means}
\usage{
scSTMseqSimBatchEffects(sim, params, nSample)
}
\arguments{
\item{sim}{SingleCellExperiment to add batch effects to.}

\item{params}{SplatParams object with simulation parameters.}
}
\value{
SingleCellExperiment with simulated gene means.

SingleCellExperiment with simulated batch effects.
}
\description{
Simulate gene means from a gamma distribution. Also simulates outlier
expression factors. Genes with an outlier factor not equal to 1 are replaced
with the median mean expression multiplied by the outlier factor.
}
\keyword{Batch}
\keyword{Simulate}
\keyword{\code{\link{getLNormFactors}}}
\keyword{added}
\keyword{along}
\keyword{and}
\keyword{are}
\keyword{batch}
\keyword{batch.}
\keyword{each}
\keyword{effect}
\keyword{effects}
\keyword{effects.}
\keyword{factors}
\keyword{for}
\keyword{internal}
\keyword{means}
\keyword{produced}
\keyword{these}
\keyword{updated}
\keyword{using}
\keyword{with}
