% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatter-AllGenerics.R,
%   R/splatter-Params-methods.R, R/splatter-SplatParams-methods.R
\name{setParam}
\alias{setParam}
\alias{setParam,Params-method}
\alias{setParam,SplatParams-method}
\title{Set a parameter}
\usage{
setParam(object, name, value)

\S4method{setParam}{Params}(object, name, value)

\S4method{setParam}{SplatParams}(object, name, value)
}
\arguments{
\item{object}{object to set parameter in.}

\item{name}{name of the parameter to set.}

\item{value}{value to set the parameter to.}
}
\value{
Object with new parameter value.
}
\description{
Function for setting parameter values.
}
\examples{
params <- newSimpleParams()
setParam(params, "nGenes", 100)

}
