% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatter-AllGenerics.R,
%   R/splatter-Params-methods.R
\name{setParamUnchecked}
\alias{setParamUnchecked}
\alias{setParamUnchecked,Params-method}
\title{Set a parameter UNCHECKED}
\usage{
setParamUnchecked(object, name, value)

\S4method{setParamUnchecked}{Params}(object, name, value)
}
\arguments{
\item{object}{object to set parameter in.}

\item{name}{name of the parameter to set.}

\item{value}{value to set the parameter to.}
}
\value{
Object with new parameter value.
}
\description{
Function for setting parameter values. THE OUTPUT IS NOT CHECKED FOR
VALIDITY!
}
\keyword{internal}
