% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSTMseq-simulate.R
\name{splatSimDropout}
\alias{splatSimDropout}
\title{Simulate dropout}
\usage{
splatSimDropout(sim, params)
}
\arguments{
\item{sim}{SingleCellExperiment to add dropout to.}

\item{params}{SplatParams object with simulation parameters.}
}
\value{
SingleCellExperiment with simulated dropout and observed counts.
}
\description{
A logistic function is used to form a relationship between the expression
level of a gene and the probability of dropout, giving a probability for each
gene in each cell. These probabilities are used in a Bernoulli distribution
to decide which counts should be dropped.
}
\keyword{internal}
