% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSTMseq-simulate.R
\name{splatSimLibSizes}
\alias{splatSimLibSizes}
\title{Simulate library sizes}
\usage{
splatSimLibSizes(sim, params)
}
\arguments{
\item{sim}{SingleCellExperiment to add library size to.}

\item{params}{SplatParams object with simulation parameters.}
}
\value{
SingleCellExperiment with simulated library sizes.
}
\description{
Simulate expected library sizes. Typically a log-normal distribution is used
but there is also the option to use a normal distribution. In this case any
negative values are set to half the minimum non-zero value.
}
\keyword{internal}
