% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatter-splat-simulate.R
\name{splatSimulate}
\alias{splatSimulate}
\alias{splatSimulateSingle}
\alias{splatSimulateGroups}
\alias{splatSimulatePaths}
\title{Simulate single-cell data using Splatter revised function}
\usage{
splatSimulate(
  params = newSplatParams(),
  method = c("single", "groups", "paths"),
  sparsify = TRUE,
  verbose = TRUE,
  cancerCellGroup = NULL,
  true_param = NULL,
  ...
)

splatSimulateSingle(params = newSplatParams(), verbose = TRUE, ...)

splatSimulateGroups(params = newSplatParams(), verbose = TRUE, ...)

splatSimulatePaths(params = newSplatParams(), verbose = TRUE, ...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[splatter:splatSimulate]{splatter::splatSimulate()}}.}
}
\value{
A \link{SingleCellExperiment} object.
}
\description{
Wrapper around revised splatter::splatSimulate to generate synthetic
single-cell data for testing or examples.
}
