% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STMsubset.R
\name{STMsubset}
\alias{STMsubset}
\title{Subset the stm results based on sample ID or a subset of document names}
\usage{
STMsubset(stmobj, subDocName = NULL, sampleID = NULL)
}
\arguments{
\item{stmobj}{An \code{STM} object created by \code{\link{multi_stm}}}

\item{subDocName}{A vector of document names in the STM object}

\item{sampleID}{A vector of sample IDs that used in the \code{\link{multi_stm}}}
}
\description{
Subset the stm results based on sample ID or a subset of document names
}
