% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSTMseq-simulate.R
\name{splatSimBatchCellMeans}
\alias{splatSimBatchCellMeans}
\title{Simulate batch means}
\usage{
splatSimBatchCellMeans(sim, params)
}
\arguments{
\item{sim}{SingleCellExperiment to add batch means to.}

\item{params}{SplatParams object with simulation parameters.}
}
\value{
SingleCellExperiment with simulated batch means.
}
\description{
Simulate a mean for each gene in each cell incorporating batch effect
factors.
}
\keyword{internal}
