% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSTMseq-simulate.R
\name{splatSimBCVMeans}
\alias{splatSimBCVMeans}
\title{Simulate means for each gene in each cell that are adjusted to follow a
mean-variance trend using Biological Coefficient of Variation taken from
and inverse gamma distribution.}
\usage{
splatSimBCVMeans(sim, params)
}
\arguments{
\item{sim}{SingleCellExperiment to add BCV means to.}

\item{params}{SplatParams object with simulation parameters.}
}
\value{
SingleCellExperiment with simulated BCV means.
}
\description{
Simulate means for each gene in each cell that are adjusted to follow a
mean-variance trend using Biological Coefficient of Variation taken from
and inverse gamma distribution.
}
\keyword{internal}
